optilogic/requirements.txt
argparse
GitPython
requests